% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagonal.R
\name{diagonal}
\alias{diagonal}
\title{Version of diag that has more consistent behavior}
\usage{
diagonal(x, len = NA, ...)
}
\arguments{
\item{x}{A vector, matrix or array with third mode length 1}

\item{...}{parameters passed to diag}
}
\value{
if x is a vector or integer, return a matrix with x on the diagonal. 
If x is a matrix, or degenerate array, return the diagonal of x.
}
\description{
Version of diag that has more consistent behavior
}
\examples{
diagonal(c(1,3))
diagonal(matrix(1:6, 2,3))
diagonal(5)
diagonal(array(1:12, dim=c(3,4,1)))
}

