% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_var.R
\name{exp_var}
\alias{exp_var}
\title{Get the explained variance for a set of predictions}
\usage{
exp_var(obs, pred, verbose = F)
}
\arguments{
\item{obs}{data.frame, vector or matrix}

\item{pred}{data.frame, vector or matrix}
}
\value{
numeric value of the explained variance
}
\description{
Calculates 1-var(obs-pred)/var(obs). If verbose == TRUE the result is printed.
}
\examples{
exp_var(rnorm(100) + seq(0,9.9,.1),  seq(0,9.9,.1))

}

