% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_influence.R
\name{get_influence}
\alias{get_influence}
\title{Given a \code{model} object, rank the input predictors (and combinations thereof)
by thier influence on the output}
\usage{
get_influence(m, d, method = "sub", interactions = TRUE)
}
\arguments{
\item{m}{a \code{Tucker_model} or \code{CP_model} object}

\item{d}{an \code{input_data} object}
}
\description{
If \code{method} is \code{'add'} then the baseline prediction is made using just the
constant coefficients (if used) and the mean squared error (MSE) is measured between 
the baseline and predictions made with each predictor added alone (univariate analysis).
}
\details{
If \code{method} is \code{'sub'} then the baseline is made using all predictors and
MSE measured for predictions made with each predictor removed.

If \code{interactions==TRUE} then MSE for predictions made with predictors for each mode
interacting are measured
}

