% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lower_bnd_Tucker.R
\name{lower_bnd_Tucker}
\alias{lower_bnd_Tucker}
\title{Calculate the lower bound of the log likelihood for a trained Tucker model}
\usage{
lower_bnd_Tucker(m, d)
}
\arguments{
\item{m}{object of the class \code{Tucker_model}}

\item{d}{object of the class \code{input_data}}
}
\value{
Returns a numerical value (should be negative)
}
\description{
Calculate the lower bound of the log likelihood for a trained Tucker model
}
\examples{
data.params <- get_data_params(c('decomp=Tucker'))
toy <- mk_toy(data.params)
train.data <- input_data$new(mode1.X=toy$mode1.X[,-1],
                             mode2.X=toy$mode2.X[,-1],
                             mode3.X=toy$mode3.X,
                             resp=toy$resp)
model.params <- get_model_params(c('decomp=Tucker'))
toy.model <- mk_model(train.data, model.params)
toy.model$rand_init(model.params)

train(d=train.data, m=toy.model, new.iter=1, params=model.params)

lower_bnd_Tucker(toy.model, train.data)
}

