% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_model.R
\name{mk_model}
\alias{mk_model}
\title{Make a new model object}
\usage{
mk_model(d, params)
}
\arguments{
\item{d}{An \code{input_data} object. See \code{input_data}.}

\item{params}{A list of parameter values \code{get_model_params}.}
}
\value{
\code{CP_model} or \code{Tucker_model} object
}
\description{
This function is a wrapper calling Tucker_model$new()  or CP_model$new() depending
on whether \code{params$decomp=='Tucker'} or \code{params$decomp=='CP'}
}
\seealso{
\link{Tucker_model}, \link{CP_model}
}

