% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_resp.R
\name{normalize_resp}
\alias{normalize_resp}
\title{z-scale responses by params$norm.mode and store means and sds as attributes}
\usage{
normalize_resp(resp.tens, params)
}
\arguments{
\item{resp.tens}{array}

\item{params}{params object}
}
\value{
normalized array
}
\description{
z-scale responses by params$norm.mode and store means and sds as attributes
}

