% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrmse.R
\name{nrmse}
\alias{nrmse}
\title{Computes the normalized root mean squared error}
\usage{
nrmse(obs, pred)
}
\arguments{
\item{obs}{observed vector, matrix or data.frame}

\item{pred}{predicted vector, matrix or data.frame}
}
\value{
numeric value of the root mean squared error normalized to the standard deviation
of the observed data
}
\description{
Computes the normalized root mean squared error
}

