% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot reciever operating characteristic (ROC) curves for two projection (A) matrices}
\usage{
plot_roc(true, pred, main = character(0))
}
\arguments{
\item{true}{projection matrix where rows of true predictors have non-zero values}

\item{pred}{projection matrix where rows of learned predictors have larger values}

\item{main}{title of the ROC plot}
}
\description{
This is a little different than a typical ROC curve since any rows of the true matrix
that are non-zero are treated as equal true positives.
}

