% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_test_RMSE.R
\name{plot_test_RMSE}
\alias{plot_test_RMSE}
\title{Plot RMSE results from test data}
\usage{
plot_test_RMSE(test.results, ylim = "default", main = "Test RMSEs",
  baselines = c(warm = NA, m1 = NA, m2 = NA, m3 = NA, m1m2 = NA, m1m3 = NA,
  m2m3 = NA, m1m2m3 = NA))
}
\arguments{
\item{test.results}{An object created with \code{test_results}}

\item{ylim}{Limits of the y-axis (default is (0, 1.5))}

\item{main}{Main title of the plot}

\item{baselines}{named vector of baseline values to draw as dotted horizontal 
lines e.g. c('warm'=0, 'm1'=0, 'm1m2'=0, 'm1m2m3'=0)}
}
\description{
Plot RMSE results from test data
}

