% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_preds.R
\name{remove_preds}
\alias{remove_preds}
\title{Remove predictors from A and X matrices if indicated by remove.lmt}
\usage{
remove_preds(m, d, params)
}
\arguments{
\item{An}{object}
}
\value{
Laurem ipsum
}
\description{
\code{test}
}
\details{
Laurem ispsum... This is a generic function: methods can be defined for it
directly or via the \code{\link{Summary}} group generic. For this to work
properly, the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}

