% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Updates the root mean squared error for training data.  Predicting both from data
and from just the latent (H) matrices.}
\usage{
rmse(m, d, verbose = T)
}
\arguments{
\item{m}{training object}

\item{d}{data object}

\item{verbose}{Logical indicating whether to print the results (TRUE)}
}
\value{
numeric value of the explained variance
}
\description{
Updates the root mean squared error for training data.  Predicting both from data
and from just the latent (H) matrices.
}

