% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_prod.R
\name{safe_prod}
\alias{safe_prod}
\title{Takes the product of two matrices adding a column of constants if necessary to the first matrix.}
\usage{
safe_prod(A, B)
}
\arguments{
\item{A}{matrix one}

\item{B}{matrix two}
}
\value{
matrix product of A and B
}
\description{
Takes the product of two matrices adding a column of constants if necessary to the first matrix.
}

