% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_mat.R
\name{show_mat}
\alias{show_mat}
\title{Plot matrices from a model object with im_mat}
\usage{
show_mat(m, d, show.mode, scale = F)
}
\arguments{
\item{m}{model object created with mk_model}

\item{d}{input data object created with get_input_data}

\item{show.mode}{vector of modes whose projection and latent matrices are 
to be displayed}

\item{scale}{Logical should the columns of matrices be scaled}
}
\description{
Plot matrices from a model object with im_mat
}

