% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Get test predictions for a 3D BaTFLED model.}
\usage{
test(d, m, ...)
}
\arguments{
\item{d}{object of the class \code{input_data} created with input_data()}

\item{m}{object of the class \code{CP_model} or \code{Tucker_model} created with mk_model()}

\item{...}{extra parameters passed to test_CP or test_Tucker}
}
\value{
An array of predicted responses the same size as \code{m$resp}.
}
\description{
This is just a wrapper that calls test_CP or test_Tucker depending on the type of model provided.
}
\examples{
value

}

