% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_Tucker.R
\name{train_Tucker}
\alias{train_Tucker}
\title{Train a Tucker model using BaTFLED algorthm}
\usage{
train_Tucker(d, m, new.iter = 1, params)
}
\arguments{
\item{d}{an input data object created with \code{input_data}}

\item{m}{a \code{Tucker_model} object created with \code{mk_model}}

\item{params}{List of parameters created with \code{get_model_params()}}
}
\description{
Model objects are updated in place to avoid memory issues. Nothing is returned.
}
\examples{
data.params <- get_data_params(c('decomp=Tucker'))
toy <- mk_toy(data.params)
train.data <- input_data$new(mode1.X=toy$mode1.X[,-1],
                             mode2.X=toy$mode2.X[,-1],
                             mode3.X=toy$mode3.X,
                             resp=toy$resp)
model.params <- get_model_params(c('decomp=Tucker'))
toy.model <- mk_model(train.data, model.params)
toy.model$rand_init(model.params)

train(d=train.data, m=toy.model, new.iter=1, params=model.params)
}

