% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_model.R
\docType{class}
\name{CP_model}
\alias{CP_model}
\title{BaTFLED model object for 3-D response tensor with CP decomposition.}
\format{An \code{\link{R6Class}} generator object}
\usage{
CP_model
}
\arguments{
\item{data}{An \code{input_data} object. See \code{input_data}.}

\item{params}{A list of parameter values. See \code{get_model_params}.}
}
\description{
\code{CP_model} objects are 'R6' objects so that their values can be updated in place. 
The object is treated like an environment and components are accessed using the \code{$} 
operator. 
When creating a new CP_model object it will be populated with default values and empty matrices.
To initialize a \code{CP_model} call the \code{initialize()} method.
}
\section{Methods}{

\describe{
  \item{\code{new(data, params)}}{Creates a new \code{CP_model} object with
  matrices sized accoring to the matrices in \code{data}.}
  \item{\code{rand_init(params)}}{Initializes the \code{CP_model} with
  random values accoring to \code{params}.}
}
}
\examples{
args <- c('decomp=CP')
params <- get_model_params(args)
data <- input_data$new(mode1.X=matrix(rnorm(20),4,5),
                       mode2.X=matrix(rnorm(30),3,10),
                       mode3.X=matrix(NA, 5, 0),
                       resp=array(rnorm(60), dim=c(4,3,5)))
toy <- CP_model$new(data, params)
toy$rand_init(params)
im_mat(toy$mode1.A.mean)
}
\seealso{
\code{get_model_params}, \code{input_data}, \code{Tucker_model}
}
\keyword{data}

