% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_resp.R
\name{normalize_resp}
\alias{normalize_resp}
\title{z-scale responses by params$norm.mode and store means and sds as attributes}
\usage{
normalize_resp(resp.tens, params)
}
\arguments{
\item{An}{object of the class \code{class(model)[1]}}

\item{na.rm}{A logical scalar. Should missing values (including NaN) be
removed?}
}
\value{
If all inputs are integer and logical, then the output will be an
  integer. If integer overflow
  \url{http://en.wikipedia.org/wiki/Integer_overflow} occurs, the output will
  be NA with a warning. Otherwise it will be a length-one numeric or complex
  vector.

  Zero-length vectors have sum 0 by definition. See
  \url{http://en.wikipedia.org/wiki/Empty_sum} for more details.
}
\description{
\code{test}
}
\details{
Laurem ispsum... This is a generic function: methods can be defined for it
directly or via the \code{\link{Summary}} group generic. For this to work
properly, the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}
\examples{
sum(1:10)
sum(1:5, 6:10)
sum(F, F, F, T, T)

sum(.Machine$integer.max, 1L)
sum(.Machine$integer.max, 1)

\dontrun{
sum("a")
}
}

