% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_test_cor.R
\name{plot_test_cor}
\alias{plot_test_cor}
\title{Plot correlation results from test data}
\usage{
plot_test_cor(test.results, ylim = "default", mean = F, main = NA,
  method = "pearson")
}
\arguments{
\item{test.results}{results \code{generated with test_results}}

\item{ylim}{limits for the y-axis (NA)}

\item{mean}{Logical indicating whether to plot horizontal lines for mean predictions}

\item{main}{Main title of the plot}

\item{method}{Either 'pearson' or 'spearman' correlations}
}
\description{
Plot correlation results from test data
}

