% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_test_exp_var.R
\name{plot_test_exp_var}
\alias{plot_test_exp_var}
\title{Plot explained variance results from test data}
\usage{
plot_test_exp_var(test.results, ylim = NA, mean = F, main = NA)
}
\arguments{
\item{test.results}{an object generated with \code{test_results}}

\item{ylim}{Limits of the y-axis.}

\item{mean}{Logical to plot horizontal lines for mean prediction (FALSE)}

\item{main}{Main title of the plot}
}
\description{
Plot explained variance results from test data
}

