% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_test_RMSE.R
\name{plot_test_RMSE}
\alias{plot_test_RMSE}
\title{Plot RMSE results from test data}
\usage{
plot_test_RMSE(test.results, ylim = NA, mean = F, main = "Test RMSEs")
}
\arguments{
\item{test.results}{An object created with \code{test_results}}

\item{ylim}{Limits of the y-axis}

\item{mean}{Logical to print horizontal lines for mean predictions}

\item{main}{Main title of the plot}
}
\description{
Plot RMSE results from test data
}

