% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rot.R
\name{rot}
\alias{rot}
\title{Rotate a matrix for printing}
\usage{
rot(m)
}
\arguments{
\item{m}{matrix}
}
\value{
matrix that has been transposed and the columns reversed
}
\description{
Rotates a matrix so that when view is called the rows and columns appear in the 
same order as when looking at the matrix with print
}
\examples{

# Normally image shows a matrix with the first entry in the bottom left
# With rot the image is shown in the same order as print
}

